Fix fallout from rebasing onto the settings module move
Master moved useSetting from metabase/common/hooks to metabase/settings,
converted files this branch touched in ways auto-merge couldn't see, and
added a mock value for the now-non-null session-cookies setting:

- repoint useSetting imports to metabase/settings across call sites the
  rebased commits introduced or rewrote
- convert DocumentCommandSuggestion (added on master) off the deleted
  useMetabotName hook
- update the session-cookies mock from null to false to match the
  tightened type
