+++ enterprise/frontend/src/metabase-enterprise/transforms-python/upsells/PythonTransformsUpsellModal/PythonTransformsUpsellModal.tsx
@@ -3,10 +3,11 @@ import { t } from "ttag";
 
 import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper";
 import { trackUpsellViewed } from "metabase/common/components/upsells/components/analytics";
-import { useSetting, useStoreUrl } from "metabase/common/hooks";
+import { useStoreUrl } from "metabase/common/hooks";
 import { useSelector } from "metabase/redux";
 import { getStoreUsers } from "metabase/selectors/store-users";
 import { getUserIsAdmin } from "metabase/selectors/user";
+import { useSetting } from "metabase/settings";
 import {
   Button,
   Center,
