+++ frontend/src/metabase/auth/components/PasswordPanel/PasswordPanel.tsx
@@ -2,11 +2,11 @@ import { useCallback, useState } from "react";
 import { t } from "ttag";
 
 import type { MfaChallengeResponse } from "metabase/api/session";
-import { useSetting } from "metabase/common/hooks";
 import { PLUGIN_MULTI_FACTOR_AUTH } from "metabase/plugins";
 import { useDispatch, useSelector } from "metabase/redux";
 import type { LoginData } from "metabase/redux/auth";
 import { login } from "metabase/redux/auth";
+import { useSetting } from "metabase/settings";
 
 import { getExternalAuthProviders } from "../../selectors";
 import { AuthTextLink } from "../AuthButton";
