+++ frontend/src/metabase/databases/components/DatabaseForm/DatabaseForm.tsx
@@ -1,8 +1,8 @@
 import { useCallback, useMemo } from "react";
 import _ from "underscore";
 
-import { useSetting } from "metabase/common/hooks";
 import { Form, FormProvider } from "metabase/forms";
+import { useSetting } from "metabase/settings";
 import type { DatabaseData } from "metabase-types/api";
 
 import type { FormLocation } from "../../types";
