+++ frontend/src/metabase/metabot/components/editor-extensions/MetabotEmbed/MetabotEmbed.tsx
@@ -12,7 +12,6 @@ import { useLatest } from "react-use";
 import { t } from "ttag";
 
 import { useLazyMetabotGenerateContentQuery } from "metabase/api";
-import { useSetting } from "metabase/common/hooks";
 import CS from "metabase/css/core/index.css";
 import { MetabotIcon } from "metabase/metabot/components/MetabotIcon";
 import { useUserMetabotPermissions } from "metabase/metabot/hooks";
@@ -20,6 +19,7 @@ import { useDispatch, useSelector } from "metabase/redux";
 import { useEditorHost } from "metabase/rich_text_editing/tiptap/EditorHost";
 import { buildDraftCard } from "metabase/rich_text_editing/tiptap/extensions/shared/draft-card";
 import { wrapCardEmbed } from "metabase/rich_text_editing/tiptap/extensions/shared/layout";
+import { useSetting } from "metabase/settings";
 import { Box, Button, Flex, Icon, Text, Tooltip } from "metabase/ui";
 import type { MetabotGenerateContentRequest } from "metabase-types/api";
