+++ frontend/src/metabase/metabot/components/editor-extensions/MetabotMention/MetabotSuggestion.tsx
@@ -15,11 +15,11 @@ import type {
   Section,
 } from "metabase/common/components/AccordionList/types";
 import { searchFilter } from "metabase/common/components/AccordionList/utils";
-import { useSetting } from "metabase/common/hooks";
 import type { MenuItem } from "metabase/rich_text_editing/tiptap/extensions/shared/MenuComponents";
 import { MenuItemComponent } from "metabase/rich_text_editing/tiptap/extensions/shared/MenuComponents";
 import { SuggestionPaper } from "metabase/rich_text_editing/tiptap/extensions/shared/SuggestionPaper";
 import { useBuildDbMenuItems } from "metabase/rich_text_editing/tiptap/extensions/shared/suggestionHooks";
+import { useSetting } from "metabase/settings";
 import { Box, Group, Loader, Text } from "metabase/ui";
 import type { Database, SearchResult } from "metabase-types/api";
